requirements.txt
Flask>=2.0.3
Werkzeug>=2.0
flask-pymongo==2.3.0

bcrypt==3.2.0
google-cloud-vision==2.5.0
shazamio==0.6.0
